Start command work, more manager implementation

+++ common/src/main/java/net/william278/husktowns/claim/TownClaim.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 public record TownClaim(@NotNull Town town, @NotNull Claim claim) {
 
-    public static Optional<TownClaim> from(@NotNull Map.Entry<Long, Claim> entry, @NotNull HuskTowns plugin) {
+    public static Optional<TownClaim> from(@NotNull Map.Entry<Integer, Claim> entry, @NotNull HuskTowns plugin) {
         return plugin.findTown(entry.getKey()).map(town -> new TownClaim(town, entry.getValue()));
     }
 }
